Rename *Node.ts to *TreeItem.ts
