+++ ui/src/index.ts
@@ -15,13 +15,13 @@ export * from './extensionUserAgent';
 export * from './parseError';
 export * from './TestAzureAccount';
 export * from './TestUserInput';
-export * from './treeDataProvider/AzureNode';
-export * from './treeDataProvider/AzureParentNode';
+export * from './treeDataProvider/AzureTreeItem';
+export * from './treeDataProvider/AzureParentTreeItem';
 export * from './treeDataProvider/AzureTreeDataProvider';
 export * from './treeDataProvider/createTreeItemsWithErrorHandling';
 export * from './treeDataProvider/GenericTreeItem';
-export * from './treeDataProvider/RootNode';
-export * from './treeDataProvider/SubscriptionNode';
+export * from './treeDataProvider/RootTreeItem';
+export * from './treeDataProvider/SubscriptionTreeItem';
 export * from './wizard/AzureWizard';
 export * from './wizard/AzureWizardPromptStep';
 export * from './wizard/AzureWizardExecuteStep';
